calc ratio

+++ src/StudentNetworkSimulator.java
@@ -174,8 +174,8 @@ public class StudentNetworkSimulator extends NetworkSimulator
         System.out.println("Number of ACK packets sent by B:" + b.getCountACK());
         System.out.println("Number of corrupted packets:" + getnCorrupt());
         System.out.println(getnToLayer3());
-        System.out.println("Ratio of lost packets:" + getnLost() / getnToLayer3());
-        System.out.println("Ratio of corrupted packets:" + getnCorrupt() / getnToLayer3());
+        System.out.println("Ratio of lost packets:" + (double) getnLost() / (double) getnToLayer3());
+        System.out.println("Ratio of corrupted packets:" + (double) getnCorrupt() / (double) getnToLayer3());
         System.out.println("Average RTT:" + "<YourVariableHere>");
         System.out.println("Average communication time:" + "<YourVariableHere>");
         System.out.println("==================================================");
